ICMSLST-683 Refactor case management code (prepare response).
This also introduces ApplicationBase, a common baseclass for
Import/ExportApplication.
