+++ web/domains/case/_import/forms.py
@@ -78,25 +78,6 @@ class CreateWoodQuotaApplicationForm(CreateImportApplicationForm):
         return cleaned_data
 
 
-class ResponsePreparationForm(forms.ModelForm):
-    class Meta:
-        model = models.ImportApplication
-        fields = ("decision", "refuse_reason")
-        widgets = {"refuse_reason": forms.Textarea}
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields["decision"].required = True
-
-    def clean(self):
-        cleaned_data = super().clean()
-        if (
-            cleaned_data.get("decision") == models.ImportApplication.REFUSE
-            and cleaned_data.get("refuse_reason") == ""
-        ):
-            self.add_error("response", "This field is required when the Application is refused")
-
-
 class CoverLetterForm(forms.ModelForm):
     class Meta:
         model = models.ImportApplication
