+++ web/domains/case/_import/sanctions/views.py
@@ -150,7 +150,10 @@ def edit_goods_licence(request: HttpRequest, application_pk: int, goods_pk: int)
         )
 
         form_class = GoodsSanctionsLicenceForm
-        success_url = reverse("import:prepare-response", kwargs={"application_pk": application.pk})
+        success_url = reverse(
+            "case:prepare-response",
+            kwargs={"application_pk": application.pk, "case_type": "import"},
+        )
         template_name = "web/domains/case/import/manage/edit-goods-licence.html"
 
         return _edit_goods(
