+++ web/domains/case/_import/wood/views.py
@@ -341,10 +341,15 @@ def edit_goods(request: HttpRequest, pk: int) -> HttpResponse:
 
         if request.POST:
             form = GoodsWoodQuotaLicenceForm(request.POST, instance=application)
+
             if form.is_valid():
                 form.save()
+
                 return redirect(
-                    reverse("import:prepare-response", kwargs={"application_pk": application.pk})
+                    reverse(
+                        "case:prepare-response",
+                        kwargs={"application_pk": application.pk, "case_type": "import"},
+                    )
                 )
         else:
             form = GoodsWoodQuotaLicenceForm(instance=application)
