+++ web/domains/case/urls.py
@@ -117,16 +117,20 @@ urlpatterns = [
                 # notes (import/export)
                 path("notes/", include(note_urls)),
                 #
-                # ??? (TODO: ICMSLST-669)
-                # path("prepare-response/", views.prepare_response, name="prepare-response"),
+                # misc stuff (import/export)
+                path("prepare-response/", views.prepare_response, name="prepare-response"),
+                #
+                # TODO: ICMSLST-682
                 # path("cover-letter/", views.edit_cover_letter, name="edit-cover-letter"),
                 # path(
-                #     "cover-letter/preview/",
-                #     views.preview_cover_letter,
-                #     name="preview-cover-letter",
+                #     "cover-letter/preview/", views.preview_cover_letter, name="preview-cover-letter"
                 # ),
+                #
+                # TODO: ICMSLST-669
                 # path("licence/", views.edit_licence, name="edit-licence"),
                 # path("licence/preview/", views.preview_licence, name="preview-licence"),
+                #
+                # TODO: ICMSLST-681
                 # path("authorisation/", views.authorisation, name="authorisation"),
                 # path(
                 #     "start-authorisation/", views.start_authorisation, name="start-authorisation"
