+++ web/templates/partial/case/export/sidebar-management.html
@@ -41,7 +41,7 @@
     </a>
   </li>
   <li>
-    <a href="#">Response Preparation</a>
+    <a href="{{ icms_url('case:prepare-response', kwargs={'case_type': 'export', 'application_pk': process.pk}) }}">Response Preparation</a>
   </li>
   <li>
     <a href="#">Authorisation</a>
